6.0 Candidate 1 - Non-Entity Actions will work without entity, added 2 control actions: Break repeat, skip repeat, fixed some issues

See merge request eagles-creative/opencreative!195

+++ src/main/java/ua/mcchickenstudio/opencreative/coding/CodingBlockPlacer.java
@@ -1,6 +1,6 @@
 /*
  * OpenCreative+, Minecraft plugin.
- * (C) 2022-2025, McChicken Studio, mcchickenstudio@gmail.com
+ * (C) 2022-2026, McChicken Studio, mcchickenstudio@gmail.com
  *
  * OpenCreative+ is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -469,12 +469,12 @@ public class CodingBlockPlacer {
             case LOCATION -> {
                 if (data == null) return new ItemStack(Material.AIR);
                 double x, y ,z;
-                float yaw,pitch;
+                float yaw, pitch;
                 x = (double) data.getOrDefault("x", 0);
                 y = (double) data.getOrDefault("y", 0);
                 z = (double) data.getOrDefault("z", 0);
-                yaw = (float) data.getOrDefault("yaw", 0);
-                pitch = (float) data.getOrDefault("pitch", 0);
+                yaw = ((Double) data.getOrDefault("yaw", 0)).floatValue();
+                pitch = ((Double) data.getOrDefault("pitch", 0)).floatValue();
                 Location location = new Location(null, x, y, z, yaw, pitch);
                 setDisplayName(item, InteractListener.formatLocation(location));
                 setPersistentData(item,getCodingValueKey(),"LOCATION");
